# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kasir API",
        "description": "Dokumentasi API untuk aplikasi kasir.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base URL"
        }
    ],
    "paths": {
...
        "/toko/{toko}/staf": {
            "get": {
                "tags": [
                    "Staf Toko"
                ],
                "summary": "Daftar staf di toko",
                "operationId": "e6385e6f56dd46762bd4996b3015ae0d",
                "parameters": [
                    {
                        "name": "toko",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "kasir"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar staf berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Toko tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Staf Toko"
                ],
                "summary": "Assign staf ke toko",
                "operationId": "94f935681006430f36462d78dc362364",
                "parameters": [
                    {
                        "name": "toko",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "role"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "01kty..."
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "kasir",
                                        "enum": [
                                            "admin",
                                            "kasir"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Staf berhasil di-assign ke toko"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Toko atau user tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/toko/{toko}/staf/{user}": {
            "put": {
                "tags": [
                    "Staf Toko"
                ],
                "summary": "Update role atau status staf di toko",
                "operationId": "dc992fb673ced793f399327b5441239e",
                "parameters": [
                    {
                        "name": "toko",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "admin",
                                        "enum": [
                                            "admin",
                                            "kasir"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data staf berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Staf tidak ditemukan di toko ini"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Staf Toko"
                ],
                "summary": "Keluarkan staf dari toko",
                "operationId": "7892b0cc1f76e085f22f8a7c3937223e",
                "parameters": [
                    {
                        "name": "toko",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staf berhasil dikeluarkan dari toko"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Staf tidak ditemukan di toko ini"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan bearer token hasil login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}